Reorganize icon component and use iconNames
