+++ src/components/icon.js
@@ -1,13 +1,21 @@
 import React from 'react';
-import ArrowLeftSvg from '../../../static/edb-icons/arrow-left.svg'
-import ChevronDownSvg from '../../../static/edb-icons/chevron-down.svg'
-import ChevronRightSvg from '../../../static/edb-icons/chevron-right.svg'
-import CloseSvg from '../../../static/edb-icons/close.svg';
-import DottedBoxSvg from '../../../static/edb-icons/dotted-box.svg';
-import HamburgerSvg from '../../../static/edb-icons/hamburger.svg';
-import SearchSvg from '../../../static/edb-icons/search.svg';
+import ArrowLeftSvg from '../../static/edb-icons/arrow-left.svg'
+import ChevronDownSvg from '../../static/edb-icons/chevron-down.svg'
+import ChevronRightSvg from '../../static/edb-icons/chevron-right.svg'
+import CloseSvg from '../../static/edb-icons/close.svg';
+import DottedBoxSvg from '../../static/edb-icons/dotted-box.svg';
+import HamburgerSvg from '../../static/edb-icons/hamburger.svg';
+import SearchSvg from '../../static/edb-icons/search.svg';
 
-import iconNames from './iconNames';
+export const iconNames = {
+  ARROW_LEFT: 'arrowleft',
+  CHEVRON_DOWN: 'chevrondown',
+  CHEVRON_RIGHT: 'chevronright',
+  CLOSE: 'close',
+  DOTTED_BOX: 'dottedbox',
+  HAMBURGER: 'hamburger',
+  SEARCH: 'search',
+};
 
 function formatIconName(name) {
   return name && name.replace(/ /g, '').toLowerCase();
