--- src/components/icons/iconNames.js
@@ -1,9 +1,0 @@
-export default {
-  ARROW_LEFT: 'arrowleft',
-  CHEVRON_DOWN: 'chevrondown',
-  CHEVRON_RIGHT: 'chevronright',
-  CLOSE: 'close',
-  DOTTED_BOX: 'dottedbox',
-  HAMBURGER: 'hamburger',
-  SEARCH: 'search',
-};
