+++ src/components/layout.js
@@ -21,7 +21,7 @@ const Layout = ({ children }) => {
         components={{
           table: props => <table {...props} className="table" />,
           pre: props => (
-            <figure className="bg-dark p-3 m-0 rounded">
+            <figure className="bg-dark p-3 m-0 mb-3 rounded">
               <pre className="m-0" {...props} />
             </figure>
           ),
