+++ src/components/left-nav.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link } from 'gatsby';
-import Icon from './icons';
+import Icon, { iconNames } from './icon';
 import VersionDropdown from './version-dropdown';
 
 const baseUrl = (path, depth) => {
@@ -79,7 +79,7 @@ const Back = () => {
   return (
     <li className="ml-0 mb-3">
       <Link to="/" className="d-block py-1 align-middle small text-dark">
-        <Icon iconName="arrowleft" className="fill-black mt-n1 mr-1" width="12" height="12" />
+        <Icon iconName={iconNames.ARROW_LEFT} className="fill-black mt-n1 mr-1" width="12" height="12" />
         Back
       </Link>
     </li>
@@ -89,7 +89,7 @@ const Back = () => {
 const SectionHeading = ({ newList, path }) => {
   return (
     <li className="ml-0 mb-4 d-flex align-items-center">
-      <Icon iconName="dottedbox" className="opacity-2 mr-2" width="48" height="48" />
+      <Icon iconName={iconNames.DOTTED_BOX} className="opacity-2 mr-2" width="48" height="48" />
       <Link
         to={newList[0].path}
         className="d-block py-1 align-middle balance-text h5 m-0 text-dark"
@@ -103,7 +103,7 @@ const SectionHeading = ({ newList, path }) => {
 const SectionHeadingWithVersions = ({ newList, path, versionArray }) => {
   return (
     <li className="ml-0 mb-4 d-flex align-items-center">
-      <Icon iconName="dottedbox" className="opacity-2 mr-2" width="90" height="90" />
+      <Icon iconName={iconNames.DOTTED_BOX} className="opacity-2 mr-2" width="90" height="90" />
       <div className="rightsidenoclass">
         <Link
           to={newList[0].path}
